Removed Model interface, renamed AbstractModel to Model
