--- src/main/java/api/config/HerokuConfiguration.java
@@ -1,25 +1,0 @@
-package api.config;
-
-import org.apache.tomcat.jdbc.pool.DataSource;
-import org.flywaydb.core.Flyway;
-import org.springframework.boot.autoconfigure.AutoConfigureBefore;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.context.annotation.*;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-
-/**
- * Created by Vileven on 26.03.17.
- */
-@Profile("heroku")
-@ComponentScan
-@Configuration
-@EnableAutoConfiguration
-public class HerokuConfiguration {
-
-}
