+++ src/main/java/api/repository/AbstractBaseDAO.java
@@ -8,9 +8,7 @@ import javax.persistence.PersistenceContext;
 import javax.persistence.PersistenceUnit;
 import java.io.Serializable;
 
-/**
- * Created by Vileven on 27.03.17.
- */
+
 public abstract class AbstractBaseDAO<T extends Model<ID>, ID extends Serializable> implements BaseDAO<T, ID> {
     @PersistenceContext
     protected EntityManager entityManager;
